requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
psycopg2
requests
